# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Character Generator API
  description: ゲームキャラクター設定サポートツールのREST API
  version: 1.0.0
servers:
  - url: https://{apiId}.execute-api.ap-northeast-1.amazonaws.com/prod
    description: Production / Development
    variables:
      apiId:
        default: xxxxxxxxxx
security:
  - CognitoUserPool: []
components:
  securitySchemes:
    CognitoUserPool:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Amazon Cognito が発行する JWT トークン
  schemas:
...
    Relationship:
      type: object
      properties:
        relationshipId:
          type: string
          example: 01HZ9012ABCDEFGHJKMNPQRS
        characterIdA:
          type: string
        characterIdB:
          type: string
        relationshipType:
          type: string
          enum: [仲間, ライバル, 師弟, 恋人, 家族, 敵対]
        description:
          type: string
      required: [relationshipId, characterIdA, characterIdB, relationshipType]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - MAX_CHARACTERS_EXCEEDED
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - BEDROCK_UNAVAILABLE
            message:
              type: string
          required: [code, message]
  responses:
    Unauthorized:
      description: 認証トークンなし・無効
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 指定リソースが存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /projects/{projectId}/relationships:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string

    post:
      summary: 関係性作成（A→B と B→A を対称保存）
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                characterIdA:
                  type: string
                characterIdB:
                  type: string
                relationshipType:
                  type: string
                  enum: [仲間, ライバル, 師弟, 恋人, 家族, 敵対]
                description:
                  type: string
              required: [characterIdA, characterIdB, relationshipType]
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      summary: 関係性一覧取得
      operationId: listRelationships
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Relationship'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/relationships/{relationshipId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: relationshipId
        in: path
        required: true
        schema:
          type: string

    delete:
      summary: 関係性削除（A→B と B→A を両方削除）
      operationId: deleteRelationship
      responses:
        '204':
          description: 削除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/relationships/{relationshipId}/regenerate:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: relationshipId
        in: path
        required: true
        schema:
          type: string

    post:
      summary: 関係性再生成（既存を削除してBedrockで再生成）
      operationId: regenerateRelationship
      responses:
        '200':
          description: 再生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relationship'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Bedrock 呼び出し失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'